Split src/util/apps.js
into appLoader.js and decoration.js (which only contains legacy code).
