+++ src/app/components/__tests__/SerialPortSelector-test.jsx
@@ -45,7 +45,7 @@ jest.mock('react-bootstrap', () => ({
 }));
 
 // Do not decorate components
-jest.mock('../../../util/apps', () => ({
+jest.mock('../../../util/decoration', () => ({
     decorate: component => component,
 }));
