+++ src/app/containers/AppReloadDialogContainer.js
@@ -37,7 +37,7 @@
 import { remote } from 'electron';
 import AppReloadDialog from '../components/AppReloadDialog';
 import * as AppReloadDialogActions from '../actions/appReloadDialogActions';
-import { connect } from '../../util/apps';
+import { connect } from '../../util/decoration';
 
 function mapStateToProps(state) {
     const { appReloadDialog } = state.core;
