+++ src/shared/index.js
@@ -39,4 +39,4 @@ export { default as NavBar } from './App/NavBar';
 export { default as DeviceSelector } from './Device/DeviceSelector';
 
 export { default as coreReducer } from './coreReducer';
-export { getAppFile } from '../util/apps';
+export { getAppFile } from '../util/appLoader';
